Moved the data generation fixture for the tests back to conftest.py to be reused.
